+++ design-system/components/MediaContainer.tsx
@@ -66,7 +66,7 @@ export function MediaContainer({ content, theme, reveal, from }) {
 
   return (
     <SharePropsWithChildren selectedTheme={selectedTheme}>
-      <Wrap animate={anim} transition={selectedTheme.transitions.short}>
+      <Wrap animate={anim} transition={selectedTheme.transitions.long}>
         {content && content.length ? content : "add Content"}
       </Wrap>
     </SharePropsWithChildren>
